create internal monitors for all monitor services, as config reader and struct
